Moved profiler init code to its own file
