+++ internal/debug/flags.go
@@ -31,7 +31,6 @@ import (
 	"github.com/fjl/memsize/memsizeui"
 	"github.com/mattn/go-colorable"
 	"github.com/mattn/go-isatty"
-	"gopkg.in/DataDog/dd-trace-go.v1/profiler"
 	"gopkg.in/urfave/cli.v1"
 )
 
@@ -92,24 +91,6 @@ var (
 		Name:  "trace",
 		Usage: "Write execution trace to the given file",
 	}
-	datadogProfilerFlag = cli.BoolFlag{
-		Name:  "datadog.profiler",
-		Usage: "Enable Datadog profiler",
-	}
-	datadogProfilerExpensiveFlag = cli.BoolFlag{
-		Name:  "datadog.profiler.expensive",
-		Usage: "Enable Datadog profiler expensive metrics",
-	}
-	datadogProfilerEnvironmentFlag = cli.StringFlag{
-		Name:  "datadog.profiler.environment",
-		Usage: "Set Datadog profiler environment",
-		Value: "",
-	}
-	datadogProfilerTagsFlag = cli.StringFlag{
-		Name:  "datadog.profiler.tags",
-		Usage: "Comma-separated tags (key:value) attached to all measurements",
-		Value: "",
-	}
 )
 
 // Flags holds all command-line flags required for debugging.
@@ -126,10 +107,6 @@ var Flags = []cli.Flag{
 	blockprofilerateFlag,
 	cpuprofileFlag,
 	traceFlag,
-	datadogProfilerFlag,
-	datadogProfilerExpensiveFlag,
-	datadogProfilerEnvironmentFlag,
-	datadogProfilerTagsFlag,
 }
 
 var glogger *log.GlogHandler
@@ -234,34 +211,6 @@ func StartPProf(address string, withMetrics bool) {
 			log.Error("Failure in running pprof server", "err", err)
 		}
 	}()
-}
-
-func StartDatadogProfiler(service string, env string, version string, tags string, expensive bool) {
-	option := profiler.WithProfileTypes(
-		profiler.CPUProfile,
-		profiler.HeapProfile,
-	)
-
-	// The profiles below are disabled by default to keep overhead
-	// low, but can be enabled as needed.
-	if expensive {
-		option = profiler.WithProfileTypes(
-			profiler.CPUProfile,
-			profiler.HeapProfile,
-			profiler.BlockProfile,
-			profiler.MutexProfile,
-			profiler.GoroutineProfile,
-		)
-	}
-	err := profiler.Start(profiler.WithService(service),
-		profiler.WithEnv(env),
-		profiler.WithVersion(version),
-		profiler.WithTags(tags),
-		option,
-	)
-	if err != nil {
-		log.Error("Error starting Datadog profiler", "err", err)
-	}
 }
 
 // Exit stops all running profiles, flushing their output to the
